fix: channel store initialization and reference
